backend/app/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-dateutil==2.9.0.post0
numpy==2.0.1
python-dotenv==1.0.0